// src/web/shared/proto/jaiabot/messages/mission.ts
export interface MissionTask {
    type?: MissionTask_TaskType | undefined;
    /** for type == DIVE */
    dive?: MissionTask_DiveParameters | undefined;
    /** for type == SURFACE_DRIFT and type == DIVE */
    surface_drift?: MissionTask_DriftParameters | undefined;
    /** for type == CONSTANT_HEADING */
    constant_heading?: MissionTask_ConstantHeadingParameters | undefined;
    start_pam?: boolean | undefined;
    /** for type == STATION_KEEP */
    station_keep?: MissionTask_StationKeepParameters | undefined;
}

export enum MissionTask_TaskType {
    NONE = "NONE",
    DIVE = "DIVE",
    STATION_KEEP = "STATION_KEEP",
    SURFACE_DRIFT = "SURFACE_DRIFT",
    CONSTANT_HEADING = "CONSTANT_HEADING",
}

export interface MissionTask_DiveParameters {
    max_depth?: number | undefined;
    depth_interval?: number | undefined;
    hold_time?: number | undefined;
    bottom_dive?: boolean | undefined;
}

export interface MissionTask_DriftParameters {
    drift_time?: number | undefined;
}

export interface MissionTask_ConstantHeadingParameters {
    constant_heading?: number | undefined;
    constant_heading_time?: number | undefined;
    constant_heading_speed?: number | undefined;
}

export interface MissionTask_StationKeepParameters {
    station_keep_time?: number | undefined;
}
// ...
export interface MissionPlan_Goal {
    name?: string | undefined;
    location?: GeographicCoordinate | undefined;
    task?: MissionTask | undefined;
    moveWptMode?: boolean | undefined;
}